Moved 'ImageCommentsUIComposer' to its own file for production
